Update step1_ecs_singlenode_install.py
I ran into this error: "ERROR [root:210] must be string or buffer, not list".

I believe this occurs when hostname begins with an IP address. To manually set the hostname to a string, I've updated line 188 to pass a string value into /etc/hostname. This overcame the error and deployed successfully.